Arreglar cosas, puse las_name, y hacer que funcione el crear usuarios con posts

+++ database/migrations/2024_10_24_180232_add_published_at_to_posts_table.php
@@ -12,7 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->date('published_at')->change();
+            // Agregar la columna `published_at` solo si no existe
+            if (!Schema::hasColumn('posts', 'published_at')) {
+                $table->date('published_at')->nullable();
+            }
         });
     }
 
@@ -22,7 +25,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->dropColumn('published_at');
+            // Eliminar la columna `published_at` solo si existe
+            if (Schema::hasColumn('posts', 'published_at')) {
+                $table->dropColumn('published_at');
+            }
         });
     }
 };
